tests/wpt/web-platform-tests/websockets/constructor/021.html

<!doctype html>
<title>WebSockets: Same sub protocol twice</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script src=../constants.js?pipe=sub></script>
<div id=log></div>
<script>
test(function() {assert_throws("SyntaxError", function(){new WebSocket("ws://certo2.oslo.osa/protocol_array",["foobar, foobar"])})});
</script>
